Overall, I thought that your code was really well made and because it covers all the course topics, the game was simple yet followed all the criteria.
Some of my changes that I made was that I made it so the questions would be shuffled and randomized everytime you start a new game as I felt that it allowed for replayability. I also added a pause timer button in order to pause the timer for those who would like the option to. Additionally, I made the window larger for better visibility. Further, I added it so that the player could see their score in real time.

Some suggestions or TODO comments that I made was possibly having sound effects everytime there was a right or wrong answer. Obviously if you wanted them to not know then this isnt really needed. Also, I made some minor changes and fixes so that if you wanted to add on to the game, it would make it easier. Besides that, this was well done.